AI_Powered_Eduction_system/requirements.txt
imbalanced_learn==0.12.3
imblearn==0.0
matplotlib==3.10.3
nltk==3.9.1
numpy==1.23.5
pandas==2.3.1
Pillow==11.3.0
plotly==6.0.1
plotly==5.24.1
scikit_learn==1.7.0
seaborn==0.13.2
streamlit==1.43.2
streamlit==1.44.1
streamlit==1.37.1
tensorflow==2.19.0
torch==2.7.1+cu118
torchmetrics==1.7.3
torchvision==0.22.1+cu118
transformers==4.53.1
xgboost==3.0.2
